simplifies builder tests dramatically
