+++ tests/Unit/Commands/Builders/TestSetup.php
@@ -21,11 +21,6 @@ class TestSetup extends \PHPUnit_Framework_TestCase
      */
     protected $builder;
 
-    public function setup()
-    {
-        $this->builder = new BaseBuilder();
-    }
-
     // To test a JSON representation of the Command passed through a mock processor (for dispatched commands)
     public function buildExpectedCommand(array $properties)
     {
